6337 | 12 files | M DumpSource2/schemas/client/CDOTA_Modifier_Elder_Titan_EarthSplitter_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Kez_RavensVeil_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Lion_FingerPunch.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Elder_Titan_EarthSplitter_Thinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Kez_RavensVeil_Thinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Lion_FingerPunch.h,  M game/bin/built_from_cl.txt,  M game/dota/bin/linuxsteamrt64/libclient_strings.txt,  M game/dota/bin/linuxsteamrt64/libserver_strings.txt,  M game/dota/pak01_dir.txt,  M game/dota/readonly_tools_asset_info.txt,  M game/dota/steam.inf
https://steamdb.info/patchnotes/17537588/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Kez_RavensVeil_Thinker.h
@@ -4,8 +4,8 @@ class CDOTA_Modifier_Kez_RavensVeil_Thinker
 	GameTime_t m_fLastThink;
 	CUtlVector< CHandle< C_BaseEntity > > m_entitiesHit;
 	CountdownTimer m_ViewerTimer;
-	int32 blast_radius;
-	int32 blast_speed;
+	float32 blast_radius;
+	float32 blast_speed;
 	int32 apply_parry_bonus;
 	float32 blind_duration;
 	float32 vuln_duration;
