+++ DumpSource2/schemas/client/CDOTA_Modifier_Lion_FingerPunch.h
@@ -1,11 +1,11 @@
 class CDOTA_Modifier_Lion_FingerPunch
 {
-	int32 punch_attack_range;
-	int32 punch_bonus_damage_base;
-	int32 punch_bonus_damage_per_stack;
-	int32 cleave_starting_width;
-	int32 cleave_ending_width;
-	int32 cleave_distance;
-	int32 cleave_damage;
+	float32 punch_attack_range;
+	float32 punch_bonus_damage_base;
+	float32 punch_bonus_damage_per_stack;
+	float32 cleave_starting_width;
+	float32 cleave_ending_width;
+	float32 cleave_distance;
+	float32 cleave_damage;
 	int32 m_iOriginalAttackCapabilities;
 };
