+++ DumpSource2/schemas/server/CDOTA_Modifier_Elder_Titan_EarthSplitter_Thinker.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Elder_Titan_EarthSplitter_Thinker
 {
-	int32 crack_width;
+	float32 crack_width;
 	int32 damage_pct;
 	float32 slow_duration;
 	float32 slow_duration_scepter;
